backend/requirements.txt
fastapi[standard]
uvicorn
easyocr
opencv-python
pillow
spacy
